commands: Move getMountedVolumes function
